+++ lib/main.dart
@@ -6,7 +6,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:flutter_styled_toast/flutter_styled_toast.dart';
+import 'package:bot_toast/bot_toast.dart';
 
 import 'l10n/l10n.dart';
 import 'components/theme_switcher.dart';
@@ -29,46 +29,42 @@ class App extends StatelessWidget {
       child: Builder(
         builder: (BuildContext context) => ThemeSwitcherWidget(
           child: Builder(
-            builder: (BuildContext context) => StyledToast(
-              duration: Duration(seconds: 5),
-              child: MaterialApp(
-                title: 'FluffyChat',
-                theme: ThemeSwitcherWidget.of(context).themeData,
-                localizationsDelegates: [
-                  AppLocalizationsDelegate(),
-                  GlobalMaterialLocalizations.delegate,
-                  GlobalWidgetsLocalizations.delegate,
-                  GlobalCupertinoLocalizations.delegate,
-                ],
-                supportedLocales: [
-                  const Locale('en'), // English
-                  const Locale('de'), // German
-                  const Locale('hu'), // Hungarian
-                  const Locale('pl'), // Polish
-                ],
-                locale: kIsWeb
-                    ? Locale(html.window.navigator.language.split("-").first)
-                    : null,
-                home: FutureBuilder<LoginState>(
-                  future: Matrix.of(context)
-                      .client
-                      .onLoginStateChanged
-                      .stream
-                      .first,
-                  builder: (context, snapshot) {
-                    if (!snapshot.hasData) {
-                      return Scaffold(
-                        body: Center(
-                          child: CircularProgressIndicator(),
-                        ),
-                      );
-                    }
-                    if (Matrix.of(context).client.isLogged()) {
-                      return ChatListView();
-                    }
-                    return HomeserverPicker();
-                  },
-                ),
+            builder: (BuildContext context) => MaterialApp(
+              title: 'FluffyChat',
+              builder: BotToastInit(),
+              navigatorObservers: [BotToastNavigatorObserver()],
+              theme: ThemeSwitcherWidget.of(context).themeData,
+              localizationsDelegates: [
+                AppLocalizationsDelegate(),
+                GlobalMaterialLocalizations.delegate,
+                GlobalWidgetsLocalizations.delegate,
+                GlobalCupertinoLocalizations.delegate,
+              ],
+              supportedLocales: [
+                const Locale('en'), // English
+                const Locale('de'), // German
+                const Locale('hu'), // Hungarian
+                const Locale('pl'), // Polish
+              ],
+              locale: kIsWeb
+                  ? Locale(html.window.navigator.language.split("-").first)
+                  : null,
+              home: FutureBuilder<LoginState>(
+                future:
+                    Matrix.of(context).client.onLoginStateChanged.stream.first,
+                builder: (context, snapshot) {
+                  if (!snapshot.hasData) {
+                    return Scaffold(
+                      body: Center(
+                        child: CircularProgressIndicator(),
+                      ),
+                    );
+                  }
+                  if (Matrix.of(context).client.isLogged()) {
+                    return ChatListView();
+                  }
+                  return HomeserverPicker();
+                },
               ),
             ),
           ),