+++ lib/views/chat.dart
@@ -15,7 +15,7 @@ import 'package:fluffychat/l10n/l10n.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_styled_toast/flutter_styled_toast.dart';
+import 'package:bot_toast/bot_toast.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:pedantic/pedantic.dart';
 
@@ -181,7 +181,7 @@ class _ChatState extends State<_Chat> {
 
   void sendFileAction(BuildContext context) async {
     if (kIsWeb) {
-      showToast(L10n.of(context).notSupportedInWeb);
+      BotToast.showText(text: L10n.of(context).notSupportedInWeb);
       return;
     }
     File file = await FilePicker.getFile();
@@ -195,7 +195,7 @@ class _ChatState extends State<_Chat> {
 
   void sendImageAction(BuildContext context) async {
     if (kIsWeb) {
-      showToast(L10n.of(context).notSupportedInWeb);
+      BotToast.showText(text: L10n.of(context).notSupportedInWeb);
       return;
     }
     File file = await ImagePicker.pickImage(
@@ -213,7 +213,7 @@ class _ChatState extends State<_Chat> {
 
   void openCameraAction(BuildContext context) async {
     if (kIsWeb) {
-      showToast(L10n.of(context).notSupportedInWeb);
+      BotToast.showText(text: L10n.of(context).notSupportedInWeb);
       return;
     }
     File file = await ImagePicker.pickImage(