+++ lib/views/chat_details.dart
@@ -13,9 +13,10 @@ import 'package:fluffychat/views/invitation_selection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_styled_toast/flutter_styled_toast.dart';
+import 'package:bot_toast/bot_toast.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:link_text/link_text.dart';
+import './settings_emotes.dart';
 
 class ChatDetails extends StatefulWidget {
   final Room room;
@@ -39,7 +40,7 @@ class _ChatDetailsState extends State<ChatDetails> {
       widget.room.setName(displayname),
     );
     if (success != false) {
-      showToast(L10n.of(context).displaynameHasBeenChanged);
+      BotToast.showText(text: L10n.of(context).displaynameHasBeenChanged);
     }
   }
 
@@ -100,7 +101,7 @@ class _ChatDetailsState extends State<ChatDetails> {
       widget.room.setDescription(displayname),
     );
     if (success != false) {
-      showToast(L10n.of(context).groupDescriptionHasBeenChanged);
+      BotToast.showText(text: L10n.of(context).groupDescriptionHasBeenChanged);
     }
   }
 
@@ -120,7 +121,7 @@ class _ChatDetailsState extends State<ChatDetails> {
       ),
     );
     if (success != false) {
-      showToast(L10n.of(context).avatarHasBeenChanged);
+      BotToast.showText(text: L10n.of(context).avatarHasBeenChanged);
     }
   }
 
@@ -171,7 +172,8 @@ class _ChatDetailsState extends State<ChatDetails> {
                             Clipboard.setData(
                               ClipboardData(text: widget.room.canonicalAlias),
                             );
-                            showToast(L10n.of(context).copiedToClipboard);
+                            BotToast.showText(
+                                text: L10n.of(context).copiedToClipboard);
                           },
                         ),
                       ChatSettingsPopupMenu(widget.room, false)
@@ -274,6 +276,22 @@ class _ChatDetailsState extends State<ChatDetails> {
                                         ? widget.room.canonicalAlias
                                         : L10n.of(context).none),
                               ),
+                            ListTile(
+                              leading: CircleAvatar(
+                                backgroundColor:
+                                    Theme.of(context).scaffoldBackgroundColor,
+                                foregroundColor: Colors.grey,
+                                child: Icon(Icons.insert_emoticon),
+                              ),
+                              title: Text(L10n.of(context).emoteSettings),
+                              onTap: () async =>
+                                  await Navigator.of(context).push(
+                                AppRoute.defaultRoute(
+                                  context,
+                                  EmotesSettingsView(room: widget.room),
+                                ),
+                              ),
+                            ),
                             PopupMenuButton(
                               child: ListTile(
                                 leading: CircleAvatar(